+++ motion/settings.py
@@ -374,6 +374,14 @@ By default, `VIDEO_MAX_WIDTH` is 400.
 
 """
 
+INLINE_COMMENT_COUNT = 0
+"""Controls the number of comments displayed for assets within a list
+view.
+
+By default, `INLINE_COMMENT_COUNT` is 0. Assign a non-zero (up to 50) number
+to enable this feature. Between 3 and 5 is recommended.
+"""
+
 THEME = 'motion'
 """The name of the theme to use for this Motion site.
 
@@ -386,3 +394,13 @@ By default, the ``motion`` theme is used.
 
 """
 
+USE_FACEBOOK = True
+"""Controls whether a Facebook sharing button is included in the footer of
+posts.
+
+"""
+
+USE_TWEETMEME = True
+"""Controls whether a "Retweet" button is included in the footer of posts.
+
+"""